Added more commands

+++ dante/staging.py
@@ -3,6 +3,8 @@ import os
 import sys
 from pathlib import Path
 from typing import Dict
+
+from click import decorators
 import rich
 import toml
 import click
@@ -15,10 +17,22 @@ class DANTE:
     def __init__(self) -> None:
         if not os.path.exists(self.GLOBAL_FOLDER):
             os.makedirs(self.GLOBAL_FOLDER)
-        self._staging_file = os.path.join(self.GLOBAL_FOLDER, '.dante_staging')
-        self._local_config_file = os.path.join(self.LOCAL_FOLDER, 'config')
+        self._here = self._find_dante_root()
+        self._staging_file = os.path.join(
+            self._here, 'staging'
+        ) if self._here else ''
+        self._local_config_file = os.path.join(
+            self._here, 'config'
+        ) if self._here else ''
         self._stage_status: Dict = {}
         self._local_config: Dict = {}
+
+    def check_is_repo(self) -> None:
+        if not self._find_dante_root():
+            click.echo(
+                "fatal: not a dante repository, run 'dante init' to initialise"
+            )
+            sys.exit(1)
 
     def _find_dante_root(self) -> str:
         """Locate the .dante folder within the current hierarchy
@@ -37,9 +51,7 @@ class DANTE:
             if os.path.exists(_dante_dir):
                 return os.path.abspath(_dante_dir)
             _current_dir = Path(_current_dir).parent
-        raise FileNotFoundError(
-            "fatal: not an dante repository, run 'dante init' to initialise"
-        )
+        return ''
 
     def __enter__(self) -> None:
         if os.path.exists(self._staging_file):
@@ -57,6 +69,7 @@ class DANTE:
         return self
 
     def change_staging_state(self, file_to_stage: str, stage=True) -> None:
+        self.check_is_repo()
         if not os.path.exists(file_to_stage):
             click.echo(
                 f"No such file '{file_to_stage}."
@@ -68,10 +81,11 @@ class DANTE:
         )
         self._stage_status[_label] = stage
 
-    def set_configuration():
-        pass
+    def set_configuration(self):
+        self.check_is_repo()
 
     def remove_file(self, file_name: str, cached: bool = False) -> None:
+        self.check_is_repo()
         _label = os.path.relpath(
             file_name,
             os.path.dirname(self._staging_file)
@@ -84,10 +98,50 @@ class DANTE:
         if not cached:
             os.remove(file_name)
 
-    def add_remote(self, label: str, remote_url: str) -> None:
+    def add_remote(self, remote_url: str, label: str = 'origin') -> None:
+        self.check_is_repo()
+        if 'remotes' not in self._local_config:
+            self._local_config['remotes'] = {}
+        if label in self._local_config['remotes']:
+            click.echo(
+                f"error: registry remote '{label}' already exists."
+            )
+            sys.exit(1)
         self._local_config['remotes'][label] = remote_url
 
+    def remove_remote(self, label: str) -> None:
+        self.check_is_repo()
+        if 'remotes' not in self._local_config or label not in self._local_config:
+            self.fail(
+                f"No such entry '{label}' in available remotes"
+            )
+            sys.exit(1)
+        del self._local_config[label]
+
+    def modify_remote(self, label: str, url: str) -> None:
+        self.check_is_repo()
+        if 'remotes' not in self._local_config or label not in self._local_config:
+            click.echo(
+                f"No such entry '{label}' in available remotes"
+            )
+            sys.exit(1)
+        self._local_config[label] = url
+
+    def list_remotes(self, verbose: bool = False) -> None:
+        self.check_is_repo()
+        if 'remotes' not in self._local_config:
+            return
+        else:
+            _remote_print = []
+            for remote, url in self._local_config['remotes'].items():
+                _out_str = remote
+                if verbose:
+                    _out_str += f' {url}'
+                _remote_print.append(_out_str)
+            click.echo('\n'.join(_remote_print))
+
     def status(self) -> None:
+        self.check_is_repo()
         _staged = [i for i, j in self._stage_status.items() if j]
         _unstaged = [i for i, j in self._stage_status.items() if not j]
 
@@ -107,11 +161,23 @@ class DANTE:
     def initialise(self) -> None:
         """Initialise an dante repository within the current location"""
         _dante_dir = os.path.abspath(os.path.join(os.getcwd(), '.dante'))
-        os.mkdir(_dante_dir)
+        try:
+            os.mkdir(_dante_dir)
+        except FileExistsError:
+            click.echo(
+                f'fatal: dante repository is already initialised.'
+            )
+            sys.exit(1)
         click.echo(
             f'Initialised empty dante repository in {_dante_dir}'
         )
+        self._local_config_file = os.path.join(_dante_dir, 'config')
+        self._staging_file = os.path.join(_dante_dir, 'staging')
 
     def __exit__(self, *args) -> None:
+        if not os.path.exists(self.LOCAL_FOLDER):
+            return
         with open(self._staging_file, 'w') as f:
-            yaml.dump(self._stage_status, f)
+            toml.dump(self._stage_status, f)
+        with open(self._local_config_file, 'w') as f:
+            toml.dump(self._local_config, f)
